Migrated deprecated web config
