+++ src/main/java/io/suricate/monitoring/configuration/encoder/DatabasePasswordEncoder.java
@@ -21,15 +21,9 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
-/**
- * Database password encoder service
- */
 @Component
 public class DatabasePasswordEncoder {
 
-    /**
-     * Configure the password encoder
-     */
     @Bean
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder(11);
